pyproject.toml
[project]
name = "eve2cml"
version = "0.1.0"
description = "A CLI tool to convert EVE-NG XML topology data into CML2 topology data as YAML"
authors = [
    {name = "Ralph Schmieder", email = "rschmied@cisco.com"},
]
dependencies = [
    "beautifulsoup4>=4.12.3",
    "oyaml>=1.0",
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}

[project.scripts]
eve2cml = "eve2cml.main:main"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
dev = [
    "black>=24.2.0",
    "mypy>=1.8.0",
]
